# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server to get projects from github",
        "title": "ROSSETI-DRON API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/dron",
    "paths": {
...
        "/v1/module": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create module\nthat can do only admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "module"
                ],
                "summary": "create module",
                "parameters": [
                    {
                        "description": "Module info",
                        "name": "module",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/module.CreateModuleReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/module.CreateModuleResp"
                        }
                    },
                    "400": {
                        "description": "some user error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "401": {
                        "description": "not auth",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "500": {
                        "description": "internal",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    }
                }
            }
        },
        "/v1/module/submodule/{id}/test": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add submodule test\nthat can do only admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "module"
                ],
                "summary": "add submodule test",
                "parameters": [
                    {
                        "description": "SubModuleTest info",
                        "name": "submoduletest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/module.AddSubModuleTestReq"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "id of submodule",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/module.AddSubModuleTestResp"
                        }
                    },
                    "400": {
                        "description": "some user error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "401": {
                        "description": "not auth",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "404": {
                        "description": "module not found",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "500": {
                        "description": "internal",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    }
                }
            }
        },
        "/v1/module/{id}/submodule": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add submodule\nthat can do only admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "module"
                ],
                "summary": "add submodule",
                "parameters": [
                    {
                        "description": "SubModule info",
                        "name": "submodule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/module.AddSubModuleReq"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "id of module",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/module.AddSubModuleResp"
                        }
                    },
                    "400": {
                        "description": "some user error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "401": {
                        "description": "not auth",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "404": {
                        "description": "module not found",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "500": {
                        "description": "internal",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "err.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "module.AddAnswerResp": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "correct": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "module.AddQuestionResp": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/module.AddAnswerResp"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "module.AddSubModuleReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "module.AddSubModuleResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "moduleID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "module.AddSubModuleTestReq": {
            "type": "object",
            "properties": {
                "theoreticalTest": {
                    "$ref": "#/definitions/module.CreateTheoreticalTestReq"
                }
            }
        },
        "module.AddSubModuleTestResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "theoreticalTest": {
                    "$ref": "#/definitions/module.AddTheoreticalTestResp"
                }
            }
        },
        "module.AddTheoreticalTestResp": {
            "type": "object",
            "properties": {
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/module.AddQuestionResp"
                    }
                }
            }
        },
        "module.CreateAnswerReq": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "correct": {
                    "type": "boolean"
                }
            }
        },
        "module.CreateModuleReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "module.CreateModuleResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "module.CreateQuestionReq": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/module.CreateAnswerReq"
                    }
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "module.CreateTheoreticalTestReq": {
            "type": "object",
            "properties": {
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/module.CreateQuestionReq"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}